refactor: colocate @react-email/button source in react-email
Move button implementation under packages/react-email/src/components/button.
Point @react-email/button build at the shared entry and run tests from there.
Add @react-email/render for component specs, widen tsconfig for TSX, and fix
demo emails path in get-emails-directory-metadata spec.

Made-with: Cursor
